Remove unnecessary fromIntegral

+++ src/Cryptography/Wring.hs
@@ -77,7 +77,7 @@ roundEncryptFun rprime sbox xornary buf rond = i4 where
   xornrond = xorn rond
   i1 = mix3Parts buf rprime
   i2 = V.fromListN len $ map (sbox V.!) $
-    zipWith sboxInx (drop (fromIntegral rond) cycle3) (V.toList i1)
+    zipWith sboxInx (drop rond cycle3) (V.toList i1)
   i3 = rotBitcount i2 1
   i4 = V.fromListN len $ zipWith (+) (V.toList i3)
     (map (xor xornrond) (V.toList xornary))
@@ -96,7 +96,7 @@ roundDecryptFun rprime sbox xornary buf rond = i4 where
     (map (xor xornrond) (V.toList xornary))
   i2 = rotBitcount i1 (-1)
   i3 = V.fromListN len $ map (sbox V.!) $
-    zipWith sboxInx (drop (fromIntegral rond) cycle3) (V.toList i2)
+    zipWith sboxInx (drop rond cycle3) (V.toList i2)
   i4 = mix3Parts i3 rprime
 
 encryptFun :: Wring -> V.Vector Word8 -> V.Vector Word8
@@ -104,18 +104,18 @@ encryptFun wring buf = foldl' (roundEncryptFun rprime (sbox wring) xornary)
   buf rounds
   where
     len = V.length buf
-    xornary = xornArray (fromIntegral len)
+    xornary = xornArray len
     rprime = fromIntegral $ findMaxOrder (len `div` 3)
-    rounds = [0 .. (fromIntegral (nRounds len) -1)]
+    rounds = [0 .. (nRounds len) -1]
 
 decryptFun :: Wring -> V.Vector Word8 -> V.Vector Word8
 decryptFun wring buf = foldl' (roundDecryptFun rprime (invSbox wring) xornary)
   buf rounds
   where
     len = V.length buf
-    xornary = xornArray (fromIntegral len)
+    xornary = xornArray len
     rprime = fromIntegral $ findMaxOrder (len `div` 3)
-    rounds = reverse [0 .. (fromIntegral (nRounds len) -1)]
+    rounds = reverse [0 .. (nRounds len) -1]
 
 -- ST monad version modifies memory in place
 -- by int-e
